Clean up status setting code
